refactor: TownVoiceClient 조합 로직을 widget 레이어로 분리 (#108)
