+++ src/widgets/townVoiceSection/index.ts
@@ -1,0 +1,1 @@
+export { TownVoiceSection } from "./ui/TownVoiceSection";
